Move ListenerTester class in a file
